+++ R/JointModel.R
@@ -118,12 +118,10 @@ as.StanModule.JointModel <- function(object, ...) {
     # use the `decorated_render` constants `machine_double_eps`.
     stan_full <- decorated_render(.x = stan_full)
 
-    x <- merge(
+    merge(
         StanModule("base/functions.stan"),
         StanModule(stan_full)
     )
-
-    return(x)
 }
 
 
@@ -314,3 +312,14 @@ setMethod(
         ))
     }
 )
+
+
+
+#' @rdname getRandomEffectsNames
+#' @export
+getRandomEffectsNames.JointModel <- function(object, ...) {
+    if (is.null(object@longitudinal)) {
+        return(NULL)
+    }
+    getRandomEffectsNames(object@longitudinal)
+}